wx v2.7.0 incremental update - mega
git-svn-id: http://svn.osafoundation.org/chandler/trunk@9730 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ internal/wx/src/common/datstrm.cpp
@@ -1,10 +1,10 @@
 /////////////////////////////////////////////////////////////////////////////
-// Name:        datstrm.cpp
+// Name:        src/common/datstrm.cpp
 // Purpose:     Data stream classes
 // Author:      Guilhem Lavaux
 // Modified by: Mickael Gilabert
 // Created:     28/06/98
-// RCS-ID:      $Id: datstrm.cpp,v 1.50 2006/02/11 16:51:29 VZ Exp $
+// RCS-ID:      $Id: datstrm.cpp,v 1.51 2006/03/05 00:38:59 ABX Exp $
 // Copyright:   (c) Guilhem Lavaux
 // Licence:     wxWindows licence
 /////////////////////////////////////////////////////////////////////////////
@@ -296,7 +296,7 @@ void wxDataInputStream::ReadLL(wxLongLong *buffer, size_t size)
 wxLongLong wxDataInputStream::ReadLL(void)
 {
     wxLongLong ll;
-    DoReadLL(&ll, 1, m_input, m_be_order);
+    DoReadLL(&ll, (size_t)1, m_input, m_be_order);
     return ll;
 }
 #endif // wxUSE_LONGLONG
@@ -724,4 +724,3 @@ wxDataOutputStream& wxDataOutputStream::operator<<(float f)
 
 #endif
   // wxUSE_STREAMS
-
